requirements.txt
streamlit
google-generativeai
ollama
chromadb
sentence-transformers
python-dotenv

#pip install -r requirements.txt


